Rename command packages to follow convention
