+++ src/main/java/tutorlink/command/studentcommand/DeleteStudentCommand.java
@@ -1,4 +1,4 @@
-package tutorlink.command.student_command;
+package tutorlink.command.studentcommand;
 
 import tutorlink.student.Student;
