+++ src/test/java/tutorlink/command/assignmentcommand/AddAssignmentCommandTest.java
@@ -1,4 +1,4 @@
-package tutorlink.command.assignment_command;
+package tutorlink.command.assignmentcommand;
 
 import org.junit.jupiter.api.Test;
 import tutorlink.assignment.Assignment;
